+++ cgi/cancel_purchase.pl
@@ -27,7 +27,7 @@ $Invoices->execute;
 $Errs = "";
 
 unless ($Invoices->rows > 0) { $Errs .= "<li>Invoice cannot be found</li>\n"; }
-if ($Invoice[8] > 0) { $Errs .= "<li>Only Purchase Invoices created today can be cancelled</li>\n"; }
+if ($Invoice[6] > 1 && $Invoice[8] > 0) { $Errs .= "<li>Only Purchase Invoices created today can be cancelled</li>\n"; }
 
 if ($Errs) {
 	print<<EOD;
@@ -54,29 +54,33 @@ else {
 
         $Sts = $dbh->do("update invoices set invremarks=concat(invremarks,'<br/>Cancelled'),invstatus='Cancelled',invstatuscode='0',invstatusdate=now(),invtotal=0,invvat=0,invpaid=0,invpaidvat=0 where id=$FORM{id} and acct_id='$COOKIE->{ACCT}'");
 
+#  If (and only if) the purchase invoice has a non-Draft status, update nominals and coas etc
+
+	if ($Invoice[6] > 1) {
 #  5.  Subtract from customer balance (if we have a cus id)
 
-	if ($Invoice[1]) {
-		$Sts = $dbh->do("update customers set cusbalance=cusbalance - $Invoice[9] where id=$Invoice[1] and acct_id='$COOKIE->{ACCT}'");
-	}
+		if ($Invoice[1]) {
+			$Sts = $dbh->do("update customers set cusbalance=cusbalance - $Invoice[9] where id=$Invoice[1] and acct_id='$COOKIE->{ACCT}'");
+		}
 
 #  6.  Subtract from the Creditors control (1100) and Expenses Control (4000,4100,4200) coas
 
-	$Sts = $dbh->do("update coas set coabalance=coabalance - $Invoice[9] where coanominalcode='2000' and acct_id='$COOKIE->{ACCT}'");
-        $Sts = $dbh->do("insert into nominals (acct_id,link_id,nomtype,nomcode,nomamount,nomdate) values ('$COOKIE->{ACCT}',$FORM{id},'S','2000','-$Invoice[9]',now())");
+		$Sts = $dbh->do("update coas set coabalance=coabalance - $Invoice[9] where coanominalcode='2000' and acct_id='$COOKIE->{ACCT}'");
+	        $Sts = $dbh->do("insert into nominals (acct_id,link_id,nomtype,nomcode,nomamount,nomdate) values ('$COOKIE->{ACCT}',$FORM{id},'S','2000','-$Invoice[9]',now())");
 
-	$Sts = $dbh->do("update coas set coabalance=coabalance + $Invoice[3] where coanominalcode='$Invoice[2]' and acct_id='$COOKIE->{ACCT}'");
-        $Sts = $dbh->do("insert into nominals (acct_id,link_id,nomtype,nomcode,nomamount,nomdate) values ('$COOKIE->{ACCT}',$FORM{id},'S','$Invoice[2]','$Invoice[3]',now())");
+		$Sts = $dbh->do("update coas set coabalance=coabalance + $Invoice[3] where coanominalcode='$Invoice[2]' and acct_id='$COOKIE->{ACCT}'");
+	        $Sts = $dbh->do("insert into nominals (acct_id,link_id,nomtype,nomcode,nomamount,nomdate) values ('$COOKIE->{ACCT}',$FORM{id},'S','$Invoice[2]','$Invoice[3]',now())");
 
 #  7.  VAT
 
-	unless ($COOKIE->{VAT} =~ /N/) {
-		$Sts = $dbh->do("update coas set coabalance=coabalance + $Invoice[4] where coanominalcode='1400' and acct_id='$COOKIE->{ACCT}'");
-        	$Sts = $dbh->do("insert into nominals (acct_id,link_id,nomtype,nomcode,nomamount,nomdate) values ('$COOKIE->{ACCT}',$FORM{id},'S','1400','$Invoice[4]',now())");
+		unless ($COOKIE->{VAT} =~ /N/) {
+			$Sts = $dbh->do("update coas set coabalance=coabalance + $Invoice[4] where coanominalcode='1400' and acct_id='$COOKIE->{ACCT}'");
+	        	$Sts = $dbh->do("insert into nominals (acct_id,link_id,nomtype,nomcode,nomamount,nomdate) values ('$COOKIE->{ACCT}',$FORM{id},'S','1400','$Invoice[4]',now())");
 
 #  Delete from VAT accruals and deduct from comvatcontrol
 
-		$Sts = $dbh->do("delete from vataccruals where acct_id='$COOKIE->{ACCT}' and acrtxn_id=$FORM{id}");
+			$Sts = $dbh->do("delete from vataccruals where acct_id='$COOKIE->{ACCT}' and acrtxn_id=$FORM{id}");
+		}
 	}
 
 	print<<EOD;
